# back/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 문해한 하루 API
  description: 문해한 하루 API 문서입니다.
  version: 0.1.9
servers:
  - url: https://elice-kdt-ai-4th-team02.elicecoding.com
    description: Server(production)
  - url: http://localhost:5001
    description: Server(dev)
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  schemas:
...
    Subject:
      type: object
      properties:
        subject: 
          type: string
          example: "나의 일상"
        level:
          type: integer
          example: 2
        category:
          type: string
        point:
          type: integer
          example: 10
        id:
          type: string
          example: "626f9108187d6e5687442e3b"
      required:
        - subject
        - level
        - point
    Test:
      type: object
      properties:
        num: 
          type: integer
          example: 1
        question:
          type: string
          example: 문맥상 괄호 안에 가장 적합한 단어를 고르시오.
        questionType:
          type: string
          example: 객관식
        content:
          type: string
          example: "빨가면 __"
        choices:
          type: object
          example: {"1": "사과"}
      required:
        - num
        - question
        - questionType
        - choices
...
  parameters:
...
    userParams:
      name: userId
      in: path
      required: true
      description: user의 등록 id입니다. (objectId)
      schema:
        type : string
        example: 6273ed6490b074ccd0453dd9
...
    subjectParams:
      name: subjectId
      in: path
      required: true
      description: 주제 id 입니다. (objectId)
      schema:
        type : string
        example: 6273f9e181f72840e1e241ab
...
paths:
...
  /subjects:
    get:
      summary: level별 subject 조회
      parameters:
        - in: query
          name: level
          schema:
            type: integer
          description: "트레이닝 단계(level)"
              
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
      tags:
        - subjects
  /subjects/{subjectId}:
    get:
      summary: 해당 subject 조회
      parameters:
        - $ref: '#/components/parameters/subjectParams'
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
      tags:
        - subjects
  
  # test API
  /tests/evaluate:
    post:
      summary: 문해력 테스트 평가
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                1:
                  type: integer
                  example: 1
                  
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: number
                    example: 10
      tags:
        - tests
...
  /tests/{num}:
    get:
      summary: 해당 번호의 문항 조회
      parameters:
        - in: path
          required: true
          name: num
          schema:
            type: integer
          description: "문제 번호"
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
      tags:
        - tests  
  
  # result API
  /users/:userId/results:
    get:
      summary: 해당 user의 문해력 테스트 결과 조회
      parameters:
        - $ref: '#/components/parameters/userParams'
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    example: "6273ed6490b074ccd0453dd9"
                  result:
                    type: integer
                    example: 10
      tags:
        - results
      
  
  # quiz API
...